+++ src/drivers/fastfred.c
@@ -106,6 +106,47 @@ static READ_HANDLER( jumpcoas_custom_io_r )
 	return 0x00;
 }
 
+/*
+	Imago sprites DMA
+*/
+
+static UINT8 imago_sprites[0x800*3];
+static UINT16 imago_sprites_address;
+static UINT8 imago_sprites_bank = 0;
+
+static WRITE_HANDLER( imago_dma_irq_w )
+{
+	cpu_set_irq_line(0, 0, data & 1 ? ASSERT_LINE : CLEAR_LINE);
+}
+
+static WRITE_HANDLER( imago_sprites_bank_w )
+{
+	imago_sprites_bank = (data & 2) >> 1;
+}
+
+static WRITE_HANDLER( imago_sprites_dma_w )
+{
+	UINT8 *rom = (UINT8 *)memory_region(REGION_GFX2);
+	UINT8 sprites_data;
+
+	sprites_data = rom[imago_sprites_address + 0x2000*0 + imago_sprites_bank * 0x1000];
+	imago_sprites[offset + 0x800*0] = sprites_data;
+
+	sprites_data = rom[imago_sprites_address + 0x2000*1 + imago_sprites_bank * 0x1000];
+	imago_sprites[offset + 0x800*1] = sprites_data;
+
+	sprites_data = rom[imago_sprites_address + 0x2000*2 + imago_sprites_bank * 0x1000];
+	imago_sprites[offset + 0x800*2] = sprites_data;
+
+	decodechar(Machine->gfx[1], offset/32, imago_sprites, Machine->drv->gfxdecodeinfo[1].gfxlayout);
+}
+
+static READ_HANDLER( imago_sprites_offset_r )
+{
+	imago_sprites_address = offset;
+	return 0xff; /* not really used */
+}
+
 
 static MEMORY_READ_START( fastfred_readmem )
 	{ 0x0000, 0xbfff, MRA_ROM },
@@ -180,6 +221,7 @@ MEMORY_END
 
 static MEMORY_READ_START( imago_readmem )
 	{ 0x0000, 0x0fff, MRA_ROM },
+	{ 0x1000, 0x1fff, imago_sprites_offset_r },
 	{ 0x2000, 0x6fff, MRA_ROM },
 	{ 0xb000, 0xb3ff, MRA_RAM },
 	{ 0xc000, 0xc7ff, MRA_RAM },
@@ -196,6 +238,7 @@ static MEMORY_WRITE_START( imago_writemem )
 	{ 0x0000, 0x0fff, MWA_ROM },
 	{ 0x2000, 0x6fff, MWA_ROM },
 	{ 0xb000, 0xb3ff, MWA_RAM },
+	{ 0xb800, 0xbfff, imago_sprites_dma_w },
 	{ 0xc000, 0xc7ff, MWA_RAM },
 	{ 0xc800, 0xcbff, imago_fg_videoram_w, &imago_fg_videoram },
 	{ 0xd000, 0xd3ff, fastfred_videoram_w, &fastfred_videoram },
@@ -206,10 +249,12 @@ static MEMORY_WRITE_START( imago_writemem )
 	{ 0xf001, 0xf001, interrupt_enable_w },
 	{ 0xf002, 0xf002, fastfred_colorbank1_w },
 	{ 0xf003, 0xf003, fastfred_colorbank2_w },
-	{ 0xf004, 0xf004, MWA_NOP }, /* initialized with 0 then when written always 1*/
+	{ 0xf004, 0xf004, imago_dma_irq_w },
 	{ 0xf005, 0xf005, imago_charbank_w },
 	{ 0xf006, 0xf006, MWA_NOP }, /* always 0*/
 	{ 0xf007, 0xf007, MWA_NOP }, /* always 0*/
+	{ 0xf400, 0xf400, MWA_NOP },
+	{ 0xf401, 0xf401, imago_sprites_bank_w },
 	{ 0xf800, 0xf800, soundlatch_w },
 MEMORY_END
 
@@ -497,6 +542,53 @@ INPUT_PORTS_START( imago )
 	PORT_BIT( 0x10, IP_ACTIVE_HIGH, IPT_START2 )
 	PORT_BIT( 0x20, IP_ACTIVE_HIGH, IPT_UNKNOWN )
 	PORT_BIT( 0x40, IP_ACTIVE_HIGH, IPT_BUTTON1 )
+	PORT_BIT( 0x80, IP_ACTIVE_HIGH, IPT_BUTTON1 | IPF_COCKTAIL )
+
+	PORT_START
+	PORT_BIT( 0x01, IP_ACTIVE_HIGH, IPT_JOYSTICK_LEFT | IPF_8WAY )
+	PORT_BIT( 0x02, IP_ACTIVE_HIGH, IPT_JOYSTICK_RIGHT | IPF_8WAY )
+	PORT_BIT( 0x04, IP_ACTIVE_HIGH, IPT_JOYSTICK_LEFT | IPF_8WAY | IPF_COCKTAIL )
+	PORT_BIT( 0x08, IP_ACTIVE_HIGH, IPT_JOYSTICK_RIGHT | IPF_8WAY | IPF_COCKTAIL )
+	PORT_BIT( 0x10, IP_ACTIVE_HIGH, IPT_JOYSTICK_UP | IPF_8WAY )
+	PORT_BIT( 0x20, IP_ACTIVE_HIGH, IPT_JOYSTICK_DOWN | IPF_8WAY )
+	PORT_BIT( 0x40, IP_ACTIVE_HIGH, IPT_JOYSTICK_UP | IPF_8WAY | IPF_COCKTAIL )
+	PORT_BIT( 0x80, IP_ACTIVE_HIGH, IPT_JOYSTICK_DOWN | IPF_8WAY | IPF_COCKTAIL )
+
+	PORT_START
+	PORT_DIPNAME( 0x03, 0x01, DEF_STR( Lives ) )
+	PORT_DIPSETTING(    0x00, "2" )
+	PORT_DIPSETTING(    0x01, "3" )
+	PORT_DIPSETTING(    0x02, "4" )
+	PORT_DIPSETTING(    0x03, "5" )
+	PORT_DIPNAME( 0x04, 0x00, DEF_STR( Unknown ) )
+	PORT_DIPSETTING(    0x04, DEF_STR( Off ) )
+	PORT_DIPSETTING(    0x00, DEF_STR( On ) )
+	PORT_DIPNAME( 0x38, 0x00, DEF_STR( Coinage ) )
+	PORT_DIPSETTING(    0x38, DEF_STR( 5C_1C ) )
+	PORT_DIPSETTING(    0x18, DEF_STR( 4C_1C ) )
+	PORT_DIPSETTING(    0x28, DEF_STR( 3C_1C ) )
+	PORT_DIPSETTING(    0x08, DEF_STR( 2C_1C ) )
+	PORT_DIPSETTING(    0x00, DEF_STR( 1C_1C ) )
+	PORT_DIPSETTING(    0x20, DEF_STR( 1C_2C ) )
+	PORT_DIPSETTING(    0x10, DEF_STR( 1C_3C ) )
+	PORT_DIPSETTING(    0x30, DEF_STR( 1C_4C ) )	
+	PORT_DIPNAME( 0x40, 0x00, DEF_STR( Unknown ) )
+	PORT_DIPSETTING(    0x40, DEF_STR( Off ) )
+	PORT_DIPSETTING(    0x00, DEF_STR( On ) )
+	PORT_DIPNAME( 0x80, 0x00, DEF_STR( Unknown ) )
+	PORT_DIPSETTING(    0x80, DEF_STR( Off ) )
+	PORT_DIPSETTING(    0x00, DEF_STR( On ) )
+INPUT_PORTS_END
+
+INPUT_PORTS_START( imagoa )
+	PORT_START
+	PORT_BIT( 0x01, IP_ACTIVE_HIGH, IPT_UNKNOWN )
+	PORT_BIT( 0x02, IP_ACTIVE_HIGH, IPT_COIN1 )
+	PORT_BIT( 0x04, IP_ACTIVE_HIGH, IPT_UNKNOWN )
+	PORT_BIT( 0x08, IP_ACTIVE_HIGH, IPT_START1 )
+	PORT_BIT( 0x10, IP_ACTIVE_HIGH, IPT_START2 )
+	PORT_BIT( 0x20, IP_ACTIVE_HIGH, IPT_UNKNOWN )
+	PORT_BIT( 0x40, IP_ACTIVE_HIGH, IPT_BUTTON1 )
 	PORT_BIT( 0x80, IP_ACTIVE_HIGH, IPT_UNKNOWN )
 
 	PORT_START
@@ -559,6 +651,30 @@ static struct GfxLayout spritelayout =
 	32*8
 };
 
+static struct GfxLayout imago_spritelayout =
+{
+	16,16,
+	0x40,
+	3,
+	{ 0x800*8*2, 0x800*8*1, 0x800*8*0 },
+	{ 0, 1, 2, 3, 4, 5, 6, 7,
+		8*8+0, 8*8+1, 8*8+2, 8*8+3, 8*8+4, 8*8+5, 8*8+6, 8*8+7 },
+	{ 0*8, 1*8, 2*8, 3*8, 4*8, 5*8, 6*8, 7*8,
+		16*8, 17*8, 18*8, 19*8, 20*8, 21*8, 22*8, 23*8 },
+	32*8
+};
+
+static struct GfxLayout imago_char_1bpp =
+{
+	8,8,
+	RGN_FRAC(1,1),
+	1,
+	{ 0 },
+	{ 0, 1, 2, 3, 4, 5, 6, 7 },
+	{ 0*8, 1*8, 2*8, 3*8, 4*8, 5*8, 6*8, 7*8 },
+	8*8
+};
+
 static struct GfxDecodeInfo fastfred_gfxdecodeinfo[] =
 {
 	{ REGION_GFX1, 0, &charlayout,   0, 32 },
@@ -575,18 +691,19 @@ static struct GfxDecodeInfo jumpcoas_gfxdecodeinfo[] =
 
 static struct GfxDecodeInfo imago_gfxdecodeinfo[] =
 {
-	{ REGION_GFX1, 0, &charlayout,	 0, 32 },
-	{ REGION_GFX2, 0, &spritelayout, 0, 32 },
-	{ REGION_GFX3, 0, &charlayout,	 0, 32 },
+	{ REGION_GFX1, 0, &charlayout,           0,     32 },
+	{ NULL,   0xb800, &imago_spritelayout,   0,     32 },
+	{ REGION_GFX3, 0, &charlayout,           0,     32 },
+	{ REGION_GFX4, 0, &imago_char_1bpp,      0x140,  1 },
 	{ -1 } /* end of array */
 };
 
-#define CLOCK 18432000  /* The crystal is 18.432MHz */
+#define CLOCK 12432000
 
 static struct AY8910interface fastfred_ay8910_interface =
 {
 	2,             /* 2 chips */
-	CLOCK/12,       /* ? */
+	CLOCK/8,       /* 1.554 MHz */
 	{ 25, 25 },
 	{ 0 },
 	{ 0 },
@@ -597,7 +714,7 @@ static struct AY8910interface fastfred_ay8910_interface =
 static struct AY8910interface jumpcoas_ay8910_interface =
 {
 	1,             /* 1 chip */
-	CLOCK/12,       /* ? */
+	CLOCK/8,       /* 1.554 MHz */
 	{ 25 },
 	{ 0 },
 	{ 0 },
@@ -609,12 +726,12 @@ static struct AY8910interface jumpcoas_ay8910_interface =
 static MACHINE_DRIVER_START( fastfred )
 
 	/* basic machine hardware */
-	MDRV_CPU_ADD_TAG("main", Z80, CLOCK/6)     /* 3.072 MHz */
+	MDRV_CPU_ADD_TAG("main", Z80, CLOCK/4)     /* 3.108 MHz */
 	MDRV_CPU_MEMORY(fastfred_readmem,fastfred_writemem)
 	MDRV_CPU_VBLANK_INT(nmi_line_pulse,1)
 
-	MDRV_CPU_ADD_TAG("audio", Z80, CLOCK/12)
-	MDRV_CPU_FLAGS(CPU_AUDIO_CPU)    /* 1.536 MHz */
+	MDRV_CPU_ADD_TAG("audio", Z80, CLOCK/8)
+	MDRV_CPU_FLAGS(CPU_AUDIO_CPU)    /* 1.554 MHz */
 	MDRV_CPU_MEMORY(sound_readmem,sound_writemem)
 	MDRV_CPU_VBLANK_INT(nmi_line_pulse,4)
 
@@ -661,6 +778,8 @@ static MACHINE_DRIVER_START( imago )
 	MDRV_CPU_MEMORY(imago_readmem,imago_writemem)
 
 	/* video hardware */
+	MDRV_PALETTE_LENGTH(256+64+2) /* 256 for characters, 64 for the stars and 2 for the web */
+	MDRV_COLORTABLE_LENGTH(32*8+64+2)
 	MDRV_GFXDECODE(imago_gfxdecodeinfo)
 
 	MDRV_VIDEO_START(imago)
@@ -850,39 +969,74 @@ ROM_END
 
 ROM_START( imago )
 	ROM_REGION( 0x10000, REGION_CPU1, 0 )     /* 64k for main CPU */
-	ROM_LOAD( "11",    0x0000, 0x1000, CRC(3cce69b4) SHA1(e7d52e388e09e86abb597493f5807ee088cf7a40) ) 
-	ROM_CONTINUE(	   0x2000, 0x1000 )
-	ROM_LOAD( "12",    0x3000, 0x2000, CRC(8dff98c0) SHA1(e7311d9ca4544f1263e894e6d93ca52c87fc83bf) ) 
-	ROM_LOAD( "13",    0x5000, 0x2000, CRC(f0f14b4d) SHA1(92b82080575a9c95df926c404c19875ac66c2b00) ) 
+	ROM_LOAD( "imago11.82", 0x0000, 0x1000, CRC(3cce69b4) SHA1(e7d52e388e09e86abb597493f5807ee088cf7a40) )
+	ROM_CONTINUE(           0x2000, 0x1000 )
+	ROM_LOAD( "imago12.83", 0x3000, 0x2000, CRC(8dff98c0) SHA1(e7311d9ca4544f1263e894e6d93ca52c87fc83bf) )
+	ROM_LOAD( "13.bin",     0x5000, 0x2000, CRC(ae684602) SHA1(d187abbe62ee58a8190d9f428ded0feeb9484abd) )
 
-	ROM_REGION( 0x10000, REGION_CPU2, 0 )     /* 64k for audio CPU */
-	ROM_LOAD( "8",	   0x0000, 0x1000, CRC(4f77c2c9) SHA1(1e046786fbad7fb8c7c462b7bd5d80152c6b8779) ) 
-	
-	ROM_REGION( 0x3000, REGION_GFX1, ROMREGION_DISPOSE )
-	ROM_LOAD( "2",     0x0000, 0x1000, CRC(71354480) SHA1(f5f5e1cc336cae1778b7f6c744eb1bdc4226f138) ) 
-	ROM_LOAD( "3",     0x1000, 0x1000, CRC(7aba3d98) SHA1(5d058f39bf1339d523fe015b67083d44ff6a81d4) ) 
-	ROM_FILL(	       0x2000, 0x1000, 0 )
-	
-	ROM_REGION( 0x6000, REGION_GFX2, ROMREGION_DISPOSE )
-	ROM_LOAD( "4",     0x0000, 0x1000, CRC(ed987b3e) SHA1(2f88a0463b4323adb27467fb3d022144a4943793) ) 
-	ROM_LOAD( "5",     0x1000, 0x1000, CRC(77ee68ce) SHA1(a47af1bec81977d0f47463bd88e9f526fd2d6611) ) 
-	ROM_LOAD( "7",     0x2000, 0x1000, CRC(48b35190) SHA1(3a000264aad03f55fe67eed7c868acf87e804c0f) ) 
-	ROM_LOAD( "6",     0x3000, 0x1000, CRC(136990fc) SHA1(f3ecba92db25fbeb7df83c26667b7447c2d03b58) ) 
-	ROM_LOAD( "9",     0x4000, 0x1000, CRC(9efb806d) SHA1(504cc27cf071873714ec61835d9da676884fe1c8) ) 	
-	ROM_LOAD( "10",    0x5000, 0x1000, CRC(801a18d3) SHA1(f798978a47124f50be25ab4e5c6a4974d9003634) ) 
+	ROM_REGION( 0x10000, REGION_CPU2, 0 )
+	ROM_LOAD( "imago08.60", 0x0000, 0x1000, CRC(4f77c2c9) SHA1(1e046786fbad7fb8c7c462b7bd5d80152c6b8779) )
 
-	ROM_REGION( 0x3000, REGION_GFX3, ROMREGION_DISPOSE ) 
-	ROM_LOAD( "14",    0x0000, 0x1000, CRC(eded37f6) SHA1(c2ff5d4c1b001740ec4453467f879035db196a9b) ) 
-	ROM_FILL(		   0x1000, 0x2000, 0 )
+	ROM_REGION( 0x3000, REGION_GFX1, 0 )
+	ROM_LOAD( "1.bin",      0x0000, 0x1000, CRC(f80a0b69) SHA1(2b85179942586316eb61614d8697588aa9d26f9a) )
+	ROM_LOAD( "imago02.40", 0x1000, 0x1000, CRC(71354480) SHA1(f5f5e1cc336cae1778b7f6c744eb1bdc4226f138) )
+	ROM_LOAD( "3.bin",      0x2000, 0x1000, CRC(722fd625) SHA1(6f9a9f4f000cc0b251ca8d496a2ea4a708665dda) )
+
+	ROM_REGION( 0x6000, REGION_GFX2, 0 )
+	ROM_LOAD( "imago04.51", 0x0000, 0x1000, CRC(ed987b3e) SHA1(2f88a0463b4323adb27467fb3d022144a4943793) )
+	ROM_LOAD( "imago05.52", 0x1000, 0x1000, CRC(77ee68ce) SHA1(a47af1bec81977d0f47463bd88e9f526fd2d6611) )
+	ROM_LOAD( "imago07.56", 0x2000, 0x1000, CRC(48b35190) SHA1(3a000264aad03f55fe67eed7c868acf87e804c0f) )
+	ROM_LOAD( "imago06.55", 0x3000, 0x1000, CRC(136990fc) SHA1(f3ecba92db25fbeb7df83c26667b7447c2d03b58) )
+	ROM_LOAD( "imago09.64", 0x4000, 0x1000, CRC(9efb806d) SHA1(504cc27cf071873714ec61835d9da676884fe1c8) )
+	ROM_LOAD( "imago10.65", 0x5000, 0x1000, CRC(801a18d3) SHA1(f798978a47124f50be25ab4e5c6a4974d9003634) )
+
+	ROM_REGION( 0x3000, REGION_GFX3, 0 )
+	ROM_LOAD( "imago14.170", 0x0000, 0x1000, CRC(eded37f6) SHA1(c2ff5d4c1b001740ec4453467f879035db196a9b) )
+	ROM_FILL(                0x1000, 0x2000, 0x00 )
+
+	ROM_REGION( 0x1000, REGION_GFX4, 0 )
+	ROM_LOAD( "imago15.191", 0x0000, 0x1000, CRC(85fcc195) SHA1(a76f24201c037d1e6f909fb0ea4ad59b1d6ddd57) )
 
 	ROM_REGION( 0x0300, REGION_PROMS, 0 )
-	ROM_LOAD( "red",   0x0000, 0x0100, NO_DUMP )
-	ROM_LOAD( "green", 0x0100, 0x0100, NO_DUMP )
-	ROM_LOAD( "blue",  0x0200, 0x0100, NO_DUMP )
+	ROM_LOAD( "imago.96", 0x0000, 0x0100, CRC(5ba81edc) SHA1(b64ebbe054052583688cdf0f064794436c095e7e) )
+	ROM_LOAD( "imago.95", 0x0100, 0x0100, CRC(e2b7aa09) SHA1(f8edfccdd698793d9a9f423953a582b0f7b9b697) )
+	ROM_LOAD( "imago.97", 0x0200, 0x0100, CRC(e28a7f00) SHA1(05b4882c5ea5da332735866d858872bc5eeaca24) )
+ROM_END
 
-	ROM_REGION( 0x2000, REGION_USER1, 0 )	
-	ROM_LOAD( "1",      0x0000, 0x1000, CRC(b0a1fb54) SHA1(fbc746748947a7aa35a428dc862ff4ad53516d38) ) /*contains the same 2 bytes*/
-	ROM_LOAD( "15",     0x1000, 0x1000, CRC(85fcc195) SHA1(a76f24201c037d1e6f909fb0ea4ad59b1d6ddd57) ) /*unknown*/
+ROM_START( imagoa )
+	ROM_REGION( 0x10000, REGION_CPU1, 0 )     /* 64k for main CPU */
+	ROM_LOAD( "imago11.82", 0x0000, 0x1000, CRC(3cce69b4) SHA1(e7d52e388e09e86abb597493f5807ee088cf7a40) )
+	ROM_CONTINUE(	        0x2000, 0x1000 )
+	ROM_LOAD( "imago12.83", 0x3000, 0x2000, CRC(8dff98c0) SHA1(e7311d9ca4544f1263e894e6d93ca52c87fc83bf) )
+	ROM_LOAD( "imago13.84", 0x5000, 0x2000, CRC(f0f14b4d) SHA1(92b82080575a9c95df926c404c19875ac66c2b00) )
+
+	ROM_REGION( 0x10000, REGION_CPU2, 0 )     /* 64k for audio CPU */
+	ROM_LOAD( "imago08.60", 0x0000, 0x1000, CRC(4f77c2c9) SHA1(1e046786fbad7fb8c7c462b7bd5d80152c6b8779) )
+
+	ROM_REGION( 0x3000, REGION_GFX1, 0 )
+	ROM_LOAD( "imago01.39", 0x0000, 0x1000, CRC(f09fe0d4) SHA1(058af955f1758db81acd021ae3e8464c18de6bb6) )
+	ROM_LOAD( "imago02.40", 0x1000, 0x1000, CRC(71354480) SHA1(f5f5e1cc336cae1778b7f6c744eb1bdc4226f138) )
+	ROM_LOAD( "imago03.41", 0x2000, 0x1000, CRC(7aba3d98) SHA1(5d058f39bf1339d523fe015b67083d44ff6a81d4) )
+
+	ROM_REGION( 0x6000, REGION_GFX2, 0 )
+	ROM_LOAD( "imago04.51", 0x0000, 0x1000, CRC(ed987b3e) SHA1(2f88a0463b4323adb27467fb3d022144a4943793) )
+	ROM_LOAD( "imago05.52", 0x1000, 0x1000, CRC(77ee68ce) SHA1(a47af1bec81977d0f47463bd88e9f526fd2d6611) )
+	ROM_LOAD( "imago07.56", 0x2000, 0x1000, CRC(48b35190) SHA1(3a000264aad03f55fe67eed7c868acf87e804c0f) )
+	ROM_LOAD( "imago06.55", 0x3000, 0x1000, CRC(136990fc) SHA1(f3ecba92db25fbeb7df83c26667b7447c2d03b58) )
+	ROM_LOAD( "imago09.64", 0x4000, 0x1000, CRC(9efb806d) SHA1(504cc27cf071873714ec61835d9da676884fe1c8) )
+	ROM_LOAD( "imago10.65", 0x5000, 0x1000, CRC(801a18d3) SHA1(f798978a47124f50be25ab4e5c6a4974d9003634) )
+
+	ROM_REGION( 0x3000, REGION_GFX3, 0 )
+	ROM_LOAD( "imago14.170", 0x0000, 0x1000, CRC(eded37f6) SHA1(c2ff5d4c1b001740ec4453467f879035db196a9b) )
+	ROM_FILL(                0x1000, 0x2000, 0 )
+
+	ROM_REGION( 0x1000, REGION_GFX4, 0 )
+	ROM_LOAD( "imago15.191", 0x0000, 0x1000, CRC(85fcc195) SHA1(a76f24201c037d1e6f909fb0ea4ad59b1d6ddd57) )
+
+	ROM_REGION( 0x0300, REGION_PROMS, 0 )
+	ROM_LOAD( "imago.96", 0x0000, 0x0100, CRC(5ba81edc) SHA1(b64ebbe054052583688cdf0f064794436c095e7e) )
+	ROM_LOAD( "imago.95", 0x0100, 0x0100, CRC(e2b7aa09) SHA1(f8edfccdd698793d9a9f423953a582b0f7b9b697) )
+	ROM_LOAD( "imago.97", 0x0200, 0x0100, CRC(e28a7f00) SHA1(05b4882c5ea5da332735866d858872bc5eeaca24) )
 ROM_END
 
 extern int fastfred_hardware_type;
@@ -931,4 +1085,5 @@ GAME( 1982, fastfred, flyboy, fastfred, fastfred, fastfred, ROT90, "Atari", "Fas
 GAME( 1983, jumpcoas, 0,      jumpcoas, jumpcoas, jumpcoas, ROT90, "Kaneko", "Jump Coaster" )
 GAME( 1983, boggy84,  0,      jumpcoas, boggy84,  boggy84,  ROT90, "bootleg", "Boggy '84" )
 GAME( 1986, redrobin, 0,      fastfred, redrobin, flyboy,   ROT90, "Elettronolo", "Red Robin" )
-GAMEX(1983, imago,    0,      imago,	imago,	  imago,    ROT90, "Acom", "Imago", GAME_IMPERFECT_GRAPHICS | GAME_WRONG_COLORS )
+GAME( 1984, imago,    0,      imago,	imago,	  imago,    ROT90, "Acom", "Imago (cocktail set)" )
+GAME( 1983, imagoa,   imago,  imago,	imagoa,	  imago,    ROT90, "Acom", "Imago (no cocktail set)" )